move kms_test.go from cmd/ostracon/commands to privval/test_util.go
